Had a circular buffer issue with RX data on serial event. RX circular was fine, but current read pointer was stuck at zero.
Idea being serialEvent gets fired and the serialEvent has a circular buffer to store the current string in.

Then the application may be behind the actual RXed data, fairly unlikely but possible in which case to get around it I have a separate current_rx_pointer which needs to be incremented and mod with 4 (size of the circular buffer)

I may need to change this slightly in that in one serialEvent can strip off all the RX events in one go until it is 1 behind the rx_pointer. rx_pointer points to the "next" location data will be place. Hence the 1 behind for current_rx_pointer.

To check. Current works fine for slow data and a simple application, but may have an issue with complex slow program....